pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evolution"
version = "0.1.0"
description = "7x24 Private AI Agent — Secretary, Mentor, Trainer, Emotional Assistant, Intelligence Gatherer"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.30.0",
    "defusedxml>=0.7.1",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "markdown>=3.5",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["evolution*"]
exclude = ["tests*", "docs*", "data*", "logs*", "scripts*"]

[project.optional-dependencies]
mem0 = [
    "mem0ai[graph]>=1.0.5",
]
notion = [
    "notion-client>=2.2.0",
]
full = [
    "mem0ai[graph]>=1.0.5",
    "notion-client>=2.2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
# Exclude live integration tests (they make real API calls)
collect_ignore = ["tests/test_live_integration.py"]
